# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IoMT Device Management System API
  description: |-
    API cho hệ thống quản lý thiết bị y tế IoMT (Internet of Medical Things).
    
    Hệ thống cung cấp các chức năng:
    - Quản lý thiết bị y tế (Device Management)
    - Giám sát real-time qua MQTT
    - Quản lý bảo trì và bảo hành
    - Quản lý vòng đời thiết bị
    - Hệ thống cảnh báo tự động
    - Quản lý người dùng và phân quyền
    
    ## Authentication
    API sử dụng Bearer Token (JWT) để xác thực. 
    Include token trong header: `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - 300 requests/minute per API key
    - Header `X-RateLimit-Remaining` cho biết số request còn lại
    
    Links:
    - [Documentation](https://iomt.hoangphucthanh.vn/docs)
    - [Support](mailto:support@iomt.com)
  version: 1.0.0
  contact:
    email: support@iomt.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://iomt.hoangphucthanh.vn
    description: Production server
  - url: http://localhost:3005
    description: Local development server (HTTPS)
  - url: http://localhost:3006
    description: Local development server (HTTP redirect)
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Lấy danh sách người dùng
      operationId: getUsers
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: organization_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Danh sách người dùng
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      tags:
        - Users
      summary: Tạo người dùng mới
      operationId: createUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Tạo thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Users
      summary: Lấy thông tin user theo ID
      operationId: getUserById
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Thông tin user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags:
        - Users
      summary: Cập nhật thông tin user
      operationId: updateUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Cập nhật thành công
  /organizations:
    get:
      tags:
        - Organizations
      summary: Lấy danh sách tổ chức
      operationId: getOrganizations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Danh sách tổ chức
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
    post:
      tags:
        - Organizations
      summary: Tạo tổ chức mới
      operationId: createOrganization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreate'
      responses:
        '201':
          description: Tạo thành công
  /organizations/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Organizations
      summary: Lấy thông tin tổ chức theo ID
      operationId: getOrganizationById
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Thông tin tổ chức
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
    put:
      tags:
        - Organizations
      summary: Cập nhật thông tin tổ chức
      operationId: updateOrganization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpdate'
      responses:
        '200':
          description: Cập nhật thành công
    delete:
      tags:
        - Organizations
      summary: Xóa tổ chức
      operationId: deleteOrganization
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Xóa thành công
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        full_name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, manager, technician, operator]
        organization_id:
          type: string
          format: uuid
        department_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    UserCreate:
      type: object
      required:
        - username
        - password
        - email
        - organization_id
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        email:
          type: string
          format: email
        full_name:
          type: string
        organization_id:
          type: string
          format: uuid
        department_id:
          type: string
          format: uuid
    UserUpdate:
      type: object
      properties:
        full_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        department_id:
          type: string
          format: uuid
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Bệnh viện Đa khoa Trung ương
        type:
          type: string
          enum: [hospital, clinic, laboratory]
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    OrganizationCreate:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          type: string
          enum: [hospital, clinic, laboratory]
        address:
          type: string
        phone:
          type: string
    OrganizationUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [hospital, clinic, laboratory]
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
...